refactor: extract dedupe helpers
