+++ panel/__tests__/dedupe.spec.js
@@ -1,16 +1,21 @@
-require('ts-node/register');
-const { dedupeFindings } = require('../../word_addin_dev/app/assets/taskpane');
+const { dedupeFindings } = require('../../build/dedupe.js');
 
 describe('dedupeFindings', () => {
-  it('removes duplicates and invalid ranges', () => {
-    const input = [
-      { rule_id: 'r', start: 0, end: 5, snippet: 'abcde', severity: 'low' },
-      { rule_id: 'r', start: 0, end: 5, snippet: 'abcde', severity: 'high' },
-      { rule_id: 'r2', start: 10, end: 9, snippet: 'bad' },
-      { rule_id: 'r3', start: 0, end: 20000, snippet: 'xxxxx' }
-    ];
+  it('returns single entry for exact duplicates', () => {
+    const sample = { rule_id: 'r', start: 0, end: 5, snippet: 'abcde' };
+    const input = Array.from({ length: 5 }, () => ({ ...sample }));
     const out = dedupeFindings(input);
     expect(out.length).toBe(1);
-    expect(out[0].severity).toBe('high');
+  });
+
+  it('keeps findings with different ranges', () => {
+    const base = { rule_id: 'r', snippet: 'abcde' };
+    const input = [
+      { ...base, start: 0, end: 5 },
+      { ...base, start: 1, end: 6 },
+      { ...base, start: 2, end: 7 }
+    ];
+    const out = dedupeFindings(input);
+    expect(out.length).toBe(3);
   });
 });
