+++ word_addin_dev/app/assets/dedupe.ts
@@ -1,0 +1,39 @@
+import { AnalyzeFinding } from "./api-client";
+
+export function normalizeText(s: string | undefined | null): string {
+  if (!s) return "";
+  return s
+    .replace(/\r\n/g, "\n")
+    .replace(/\r/g, "\n")
+    .replace(/[ \t]+/g, " ")
+    .trim();
+}
+
+export function severityRank(s?: string): number {
+  const m = (s || "").toLowerCase();
+  return m === "high" ? 3 : m === "medium" ? 2 : 1;
+}
+
+export function dedupeFindings(findings: AnalyzeFinding[]): AnalyzeFinding[] {
+  const map = new Map<string, AnalyzeFinding>();
+  let invalid = 0, dupes = 0;
+  for (const f of findings || []) {
+    const snippet = normalizeText(f.snippet || "");
+    const start = typeof f.start === "number" ? f.start : undefined;
+    const end = typeof f.end === "number" ? f.end : (start !== undefined ? start + snippet.length : undefined);
+    if (typeof start !== "number" || typeof end !== "number" || end <= start || end - start > 10000) {
+      invalid++;
+      continue;
+    }
+    const key = `${f.rule_id || ""}|${start}|${end}|${snippet}`;
+    const ex = map.get(key);
+    if (!ex || severityRank(f.severity) > severityRank(ex.severity)) {
+      map.set(key, { ...f, snippet, start, end });
+    } else {
+      dupes++;
+    }
+  }
+  const res = Array.from(map.values());
+  console.log("panel:annotate", `dedupe dropped ${invalid} invalid, ${dupes} duplicates`);
+  return res;
+}
